Move SaxHandlerHelpers to pdaaal/utils folder
